requirements.txt
django
Pillow
pandas
matplotlib
openai
huggingface_hub
python-dotenv
numpy
requests